requirements.txt
APScheduler==3.6.3
dirsearch==0.4.3.post1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
requests-ntlm==1.2.0
telegram==0.0.1